frame for loose-db validation (#287)
